+++ tests/Unit/DataObjects/ScriptTest.php
@@ -8,13 +8,14 @@
 namespace OxidProfessionalServices\Usercentrics\Tests\Unit\DataObjects;
 
 use OxidProfessionalServices\Usercentrics\DataObject\Script;
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
 /**
  * Class ServiceTest
  * @package OxidProfessionalServices\Usercentrics\Tests\Unit
- * @covers \OxidProfessionalServices\Usercentrics\DataObject\Script
  */
+#[CoversClass(\OxidProfessionalServices\Usercentrics\DataObject\Script::class)]
 class ScriptTest extends TestCase
 {
     public function testHasPath(): void